Relocate proto utils
