+++ feature/map/src/google/kotlin/org/meshtastic/feature/map/MapView.kt
@@ -96,8 +96,8 @@ import org.meshtastic.core.model.util.metersIn
 import org.meshtastic.core.model.util.mpsToKmph
 import org.meshtastic.core.model.util.mpsToMph
 import org.meshtastic.core.model.util.toString
-import org.meshtastic.core.proto.formatPositionTime
 import org.meshtastic.core.ui.component.NodeChip
+import org.meshtastic.core.ui.util.formatPositionTime
 import org.meshtastic.feature.map.component.ClusterItemsListDialog
 import org.meshtastic.feature.map.component.CustomMapLayersSheet
 import org.meshtastic.feature.map.component.CustomTileProviderManagerSheet
